pyproject.toml
[project]
name = "workoutwise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]
